Can get the current UserSession statically
Made it easier for ViewModels to statically get the current UserSession. This is useful for times when you need the whole collection of Media & MapItems regardless of what collection you're looking at - like when adding Media to an Event (& needing to generate a list of all events) or searching over the entire MediaCollection.