Can add Media to MapItems, flexible Map VMs
Added options in MapViewModel to disable EditMode, to disable opening Previews when clicking on MapItems, and to hide the Toolbar.

Added an event into MapViewModel that is called when the user clicks on a MapItem. This lets outside users provide custom functionality, such as selecting the item in a Picker Popup.

Also, in MapViewModel, users can override the default function to generate MapItemViewModels by calling SetMapItemViewModelGenerator(). This lets users generate different MapItemViewModels than the default MapLocationViewModel and MapPathViewModel.

Added PreviewType field in MapItemViewModel, that should return the type of the Preview UserControl for the item. This lets the type of Preview be set in only the ViewModel.

With all of these changes, added a MapItem Picker Popup. The user can select items in a MediaCollection and press Add Selected to Map. A popup will show up with the Map and all of the MapItems on it. The toolbar is hidden, and the user cannot edit the items or open previews for them. When the user clicks on a MapItem, it selects that MapItem in the popup. The user can Accept the popup, and the media will be added to that MapItem.

To enable this, created PickMapItemPopupViewModel for the popup PickMapItemPopup. This ViewModel creates an instance of MapViewModel, and adds custom behavior to the MapItem click. The user can only choose one MapItem at a time, since how can Media be from two places at once?

Bug in MaintainedParentCollection: Was asking for PropertyType when the item contained a collection of the desired field. It only should do that for a singleton of the field.