Switched to DataContractSerializer
Switched to DataContractSerializer because it supports references to objects (so the entire definition of each MapItem is not repeated for every Media that belongs to it).

For 3rd party classes, the default is to serialize all of the properties. So added Serializable versions of 3rd party classes Location and LocationCollection, so the Serializable classes expose only the properties that should be serialized.