When editing, MapItem faded color & show above all.
Created SelectedPin and SelectedLine MapLayers so that the currently edited MapItems are shown above all other MapItems.

Gave MapItemViewModel a FadedColor bool, which is set by MapViewModel when items are edited. When an item is being edited, all other items are given a faded color.

This is implemented as a DependencyProperty in MapItemView that changes the MapItem color when set. I tried to get the color change to work with a DataTrigger, but the MapPolyline Stroke is not actually a DependencyProperty. The MapPolyline class contains a Shape object which has the StrokeProperty. Maybe there's a way to bind to that, but I couldn't figure it out.