Revised MapLayers to control display order
In Map, switched from MapLayers to MapItemsControl. The advantage of these is that their ItemsSource properties respond to order. So now, each MapItemsControl is linked to an ObservableCollection in Map.xaml.cs, which can be sorted. To add/remove items in each ObservableCollections, MapItem views call functions in Map.xaml.cs.

For some reason, when there are Pushpins in the MapItemsControl source, changing the location of these Pushpins does not trigger a redraw of the map, even though the Location property does actually change. So each pin layer has an LayerName_Update method which removes and readds the given pin from the list. This seems like a decently efficient way to do it, given the circumstances.

Added a generic ModelVMView class to maintain a list of ViewModels from a given list of Models. This could be a base class of MediaView, but MediaView does so much more that every class would be overridden.

Switched MapViewModel to using ModelVMView.