Better loading paths from file, popup to review these paths
Revised map file loading to handle multiple tracks within a single file. Moved this option to a separate button on the Map toolbar.

When the user chooses a file & it parses properly, a popup box will open that will list all of the paths saved in the file. There, the user can mark which path will be added to the map and the name of each path. The popup will also find if any existing paths overlap with each path, and if so, the user has a couple choices for how to deal with that (add normally, replace, merge).  The user can also open up a popup box to preview the path on a map. It will also show the existing overlapping path if it exists.

Added option to PopupViewModel to hide the Accept & Cancel buttons.

MapPath view now has two DependencyProperties that serve to override the path color and thickness. These are by default null, and if they are set by the MapPathViewModel, then the path will use those values.