MediaInfo supports editing of MapItems.
The MediaInfo pane now lets you choose a MapItem for the Media. Clicking on the MapItem will open its page, and there is a button to clear the MapItem from the Media. There is a button that will let you pick a MapItem for the Media.

Generalized MapItem pages to an abstract MapItemPageViewModel.