Added ProgressBar, improved speed of overlap detection
Added ProgressBarPopup. You give the ViewModel a work function which reports its progress, and the popup has a bar which reports that progress.

Added option to PopupWindow to disallow the user from closing the popup. This will disable the close button and prevent any other method of closing from doing anything.

When loading MapPaths from a file, paths with only one point are ignored.

Greatly improved efficency of checking MapPaths for overlaps. First, gave priority to paths likely to be "fully overlapping" so that they are checked first. Second, set a minimum size for an overlapping region to be considered a "partial overlap". If this size is 50, for example, the algorithm only has to check every 50 points instead of every point. This size is set to be either 1/8 the length of the path, or if greater than 50, 50 points. This means smaller overlaps will not get detected, but the detection process is much, much quicker. Based on the track data I'm using for this, this seems like almost all overlap cases will be caught by this.