Significant speed improvements to MapPath display
The MapPath model now will calculate simplified versions of the path for each zoom level, which are passed to the Map view depending on how its zoomed.

When a path is not in view at all, it is given to the Map at the lowest zoom level (i.e the most simple version). When the path is being edited, it is always shown at the highest zoom level (the original path).

To efficiently determine whether a path is in the map viewport, the MapPath model maintans a rectangular bounding box around the path, which is used by the MapPath view to figure out if it intersects with the map viewport. Note that this is an upright rectangle. Under some cases, a rotated rectangle probably would be more efficient.

Added a RectangleD class, because the rectangle classes only supported float and int.

ProgressBarPopupViewModel now accepts a list of objects as work function arguments, instead of just one object.

In Map.xaml.cs, having a property that exposes a delegate handler, then calling += on it, started to lag really badly after 50ish adds. Just adding directly to the handler does not have this issue.