MapLocation has trees, will hide locations too close together
MapLocations now can have a list of child MapLocations. When displaying this, the children will be shown if they are far enough apart on the map at the current zoom level, otherwise the parent will be shown. This lets users have a USA pin, for example, that will expand into a couple states when zoomed in.

Fixed bug where moving a pin wouldn't save its location changes.

Added PrepareForRemoval of MediaVMView items to clean up event handlers added in the subclass.