Choose parent for new Location, FolderTree rework
MapLocationChooser control and VM which provide a tree of all the MapLocations in a given Map and let the user select one or none. So now, users can choose a parent MapLocation when creating one.

Created an abstract FolderViewModel which represents one folder in a tree. It has a list of children, handles expansion, and provides events when the folder is clicked on.

Created an abstract FolderView that subclasses ModelVMView and provides a list of FolderViewModels. Also handles clicks on any folders in the tree.

Renamed the FolderList UserControl to FolderTree, and generalized it to work on all FolderViewModels. I think someone can probably provide DataTemplates that override those in FolderTree and provide custom implementation based on the FolderViewModel subclass.

From this, created two FolderViews to display MapLocations. There must be two, because one takes in a list of MapItems and one takes in a list of MapLocations, and ObservableCollection cannot be cast to a base type. So the one that takes MapItems is passed the Map model. And the one that takes MapLocations is passed the lists of MapLocations that are stored in the MapLocation model. If the MapLocation list of children switched to the MapItem type, then only one of these would be needed, but that doesn't seem to make so much sense to do.

Updated the Media events list to use the new FolderViewModel. This is renamed now to YearMonthFolderTree, and it is not of type FolderView because it does so much custom work with the year and month folders.

Added PostCreation method to ModelVMView for subclasses that want to access the VM object after creation.

Removed PartOfEditTree from MapLocationViewModel. Now, the freezing of certain nodes in the MapLocation tree is handled entirely in MapItemView. It has a dictionary of root notes to be frozen, and each one maps to a integer that marks the depth of the tree to display, regardless of zoom level.

Changed EditableMapItem to a MapItem instead of a MapItemViewModel. It didn't make sense to have it as a MapItemViewModel, because the MapItemView is constantly destroying and recreating ViewModels. So it makes more sense to have EditableMapItem be a model instance.

Helpful functions for dealing with MapLocation trees.