Tag drop-downs show all tags, not just local collection's
Previously, a drop-down to choose from a list of tags would only show tags from the local MediaCollection. Meaning, if you're inside of an event, you will not see any tags created outside of that event. Now, all tags from the main collection in the UserSession will be shown.

One area where this is not the case is when filtering a MediaCollection. The filter will only display tags of Media that belong to that collection.

This is also cleaner because now MediaInfoViewModel doesn't need to know about the containing collection.