Added MapLocation tree to location page
Made MapLocation parent changeable and propogates the right notifications.

MapLocation page now has a clickable tree of parents, as well as a clickable list of all direct children. Clicking on any will open the page for that MapLocation.

Also, added a button on the MapLocation page to choose the parent for the MapLocation.

Set default zoom level in MapViewModel.